serperate controller for classmates

+++ IntroToMVC/Models/Classmates.cs
@@ -8,7 +8,7 @@ namespace IntroToMVC.Models
 {
     public class Classmate
     {
-        public Classmate(string name, string hairColor, double height)
+        public Classmate(string name, string hairColor, int height)
         {
             Name = name;
             HairColor = hairColor;
@@ -22,6 +22,6 @@ namespace IntroToMVC.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string HairColor { get; set; }
-        public double Height { get; set; }
+        public int Height { get; set; }
     }
 }
